# app/Http/Requests/auth_requests.go
package requests

import (
	services "fiber-starter/app/Http/Services"

	"github.com/gofiber/fiber/v3"
)

// RegisterRequest 用户注册请求
type RegisterRequest struct {
	Name     string `json:"name" validate:"required,min=2,max=100" example:"Alice"`
	Email    string `json:"email" validate:"required,email" example:"user@example.com"`
	Password string `json:"password" validate:"required,min=8" example:"password123"`
}

func (r *RegisterRequest) BindAndValidate(c fiber.Ctx) error {
	return BindAndValidateBody(c, r)
}

func (r RegisterRequest) ToInput() services.RegisterInput {
	return services.RegisterInput{
		Name:     r.Name,
		Email:    r.Email,
		Password: r.Password,
	}
}

// VerifySignUpRequest 注册邮箱验证码请求
type VerifySignUpRequest struct {
	Email string `json:"email" validate:"required,email" example:"user@example.com"`
	Code  string `json:"code" validate:"required,len=6" example:"123456"`
}

func (r *VerifySignUpRequest) BindAndValidate(c fiber.Ctx) error {
	return BindAndValidateBody(c, r)
}

func (r VerifySignUpRequest) ToInput() services.VerifyCodeInput {
	return services.VerifyCodeInput{
		Email: r.Email,
		Code:  r.Code,
	}
}

// LoginRequest 用户登录请求
type LoginRequest struct {
	Email    string `json:"email" validate:"required,email" example:"user@example.com"`
	Password string `json:"password" validate:"required" example:"password123"`
}

func (r *LoginRequest) BindAndValidate(c fiber.Ctx) error {
	return BindAndValidateBody(c, r)
}

func (r LoginRequest) ToInput() services.LoginInput {
	return services.LoginInput{
		Email:    r.Email,
		Password: r.Password,
	}
}

// RefreshTokenRequest 刷新令牌请求
type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token" validate:"required" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."` //nolint:lll
}

func (r *RefreshTokenRequest) BindAndValidate(c fiber.Ctx) error {
	return BindAndValidateBody(c, r)
}

// ChangePasswordRequest 修改密码请求
type ChangePasswordRequest struct {
	CurrentPassword string `json:"current_password" validate:"required"`
	NewPassword     string `json:"new_password" validate:"required,min=8"`
}

func (r *ChangePasswordRequest) BindAndValidate(c fiber.Ctx) error {
	return BindAndValidateBody(c, r)
}

func (r ChangePasswordRequest) ToInput(userID int64) services.ChangePasswordInput {
	return services.ChangePasswordInput{
		UserID:          userID,
		CurrentPassword: r.CurrentPassword,
		NewPassword:     r.NewPassword,
	}
}

// ResetPasswordRequest 请求密码重置验证码
type ResetPasswordRequest struct {
	Email string `json:"email" validate:"required,email"`
}

func (r *ResetPasswordRequest) BindAndValidate(c fiber.Ctx) error {
	return BindAndValidateBody(c, r)
}

func (r ResetPasswordRequest) ToInput() services.PasswordResetRequestInput {
	return services.PasswordResetRequestInput{Email: r.Email}
}

// VerifyResetPasswordRequest 验证重置密码验证码请求
type VerifyResetPasswordRequest struct {
	Email string `json:"email" validate:"required,email"`
	Code  string `json:"code" validate:"required,len=6"`
}

func (r *VerifyResetPasswordRequest) BindAndValidate(c fiber.Ctx) error {
	return BindAndValidateBody(c, r)
}

func (r VerifyResetPasswordRequest) ToInput() services.VerifyCodeInput {
	return services.VerifyCodeInput{
		Email: r.Email,
		Code:  r.Code,
	}
}

// ConfirmResetPasswordRequest 确认重置密码请求
type ConfirmResetPasswordRequest struct {
	Token    string `json:"token" validate:"required"`
	Password string `json:"password" validate:"required,min=8"`
}

func (r *ConfirmResetPasswordRequest) BindAndValidate(c fiber.Ctx) error {
	return BindAndValidateBody(c, r)
}

func (r ConfirmResetPasswordRequest) ToInput() services.ConfirmPasswordResetInput {
	return services.ConfirmPasswordResetInput{
		Token:       r.Token,
		NewPassword: r.Password,
	}
}